+++ src/pages/home/Home.tsx
@@ -5,7 +5,8 @@ import Input from '@ui/input/Input';
 import { Icon } from '@icon/Icon';
 import RatingBottomSheet from '@ui/rating-bottom-sheet/RatingBottomSheet';
 import AgreeToTermsBottomSheet from '@shared/components/agree-to-terms/AgreeToTermsBottomSheet';
-import TimePicker, { type TimeType } from '@components/TimePicker/TimePicker';
+import TimePicker from '@components/time-picker/TimePicker';
+import type { TimeType } from '@type/time-types';
 
 const Home = () => {
   // BottomSheet states
