+++ src/shared/types/time-types.ts
@@ -1,0 +1,4 @@
+export interface TimeType {
+  hour: string;
+  minute: string;
+}
